# Conflicts:
#	src/main/kotlin/com/alirezaiyan/vokab/server/service/AuthService.kt
#	src/test/kotlin/com/alirezaiyan/vokab/server/service/UserSettingsServiceTest.kt

+++ src/main/kotlin/com/alirezaiyan/vokab/server/domain/repository/WordRushGameRepository.kt
@@ -18,11 +18,11 @@ interface WordRushGameRepository : JpaRepository<WordRushGame, Long> {
                 COUNT(*)                                                                            AS totalGames,
                 COALESCE(SUM(CASE WHEN g.completed_normally THEN 1 ELSE 0 END), 0)                 AS totalCompleted,
                 COALESCE(MAX(g.best_streak), 0)                                                     AS bestStreakEver,
-                COALESCE(AVG(CAST(g.score AS DOUBLE)), 0.0)                                         AS avgScore,
-                COALESCE(AVG(CAST(g.correct_count AS DOUBLE) / NULLIF(g.total_questions, 0) * 100), 0.0)
+                COALESCE(AVG(CAST(g.score AS DOUBLE PRECISION)), 0.0)                                  AS avgScore,
+                COALESCE(AVG(CAST(g.correct_count AS DOUBLE PRECISION) / NULLIF(g.total_questions, 0) * 100), 0.0)
                                                                                                     AS avgAccuracyPercent,
                 COALESCE(SUM(g.duration_ms), 0)                                                     AS totalTimePlayedMs,
-                COALESCE(AVG(CAST(g.avg_response_ms AS DOUBLE)), 0.0)                               AS avgResponseMs
+                COALESCE(AVG(CAST(g.avg_response_ms AS DOUBLE PRECISION)), 0.0)                     AS avgResponseMs
             FROM word_rush_games g
             WHERE g.user_id = :userId
         """,
